Renamed Django app content to contentapp. Added description field to page and shortcut models.
