--- django/content/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ContentConfig(AppConfig):
-    name = 'content'
