+++ django/contentapp/templates/page_delete.html
