+++ django/contentapp/templates/shortcut_delete.html
