+++ django/contentapp/templates/shortcut_detail.html
@@ -9,7 +9,9 @@
     <div class="row justify-content-center">
         <div class="col-6 shortcut">
             <div class="cover-img">
+                {% if object.image %}
                 <img src="{{ object.image.url }}">
+                {% endif %}
             </div>
             <div class="caption text-center">
                 <h2>{{object.title_fi}}</h2>
