+++ django/contentapp/templates/shortcut_list.html
