+++ django/contentapp/templates/shortcut_update.html
